Deleted lists from the entities. Don't need them - lists can be retrieved with queries instead.
Changed the findByManager test in EmployeeDaoTest.java. Redundancy in
creating multiple managers and also don't need to check that names match
up. Just need to change that the person is a manager.

Created ajax call to pre-populate the edit form for employees. Edit
works right now by passing the emp.id to the controller - but this makes
adding not working because it is looking for the id. Will fix for next
commit.